✨ Feat: zustand 사용하여 정렬 상태관리
